+++ src/geode/mesh/CMakeLists.txt
@@ -79,7 +79,7 @@ add_geode_library(
         "core/triangulated_surface_view.cpp"
         "core/vertex_set.cpp"
         "helpers/aabb_edged_curve_helpers.cpp"
-        "helpers/aabb_triangulated_surface_helpers.cpp"
+        "helpers/aabb_surface_helpers.cpp"
         "helpers/convert_surface_mesh.cpp"
         "helpers/convert_solid_mesh.cpp"
         "io/edged_curve_input.cpp"
@@ -161,7 +161,7 @@ add_geode_library(
         "core/triangulated_surface_view.h"
         "core/vertex_set.h"
         "helpers/aabb_edged_curve_helpers.h"
-        "helpers/aabb_triangulated_surface_helpers.h"
+        "helpers/aabb_surface_helpers.h"
         "helpers/convert_surface_mesh.h"
         "helpers/convert_solid_mesh.h"
         "io/edged_curve_input.h"
