+++ tests/mesh/test-aabb-triangulated-surface-helpers.cpp
@@ -31,7 +31,7 @@
 
 #include <geode/mesh/builder/triangulated_surface_builder.h>
 #include <geode/mesh/core/triangulated_surface.h>
-#include <geode/mesh/helpers/aabb_triangulated_surface_helpers.h>
+#include <geode/mesh/helpers/aabb_surface_helpers.h>
 
 #include <geode/geometry/aabb.h>
 #include <geode/geometry/point.h>
